fix(table): MOVE table body logic to table body component
